long

+++ manydepth/trainer_m2.py
@@ -554,9 +554,9 @@ class Trainer_Monodepth:
                     pb_y1 = y - offset1[1]
                     V = self.get_v(depth_data[b],(pa_y1,pa_x1),(pb_y1,pb_x1),k[b])
                     print(normal_data.shape)
-                    x = normal_data[b][0][y][x]
-                    y = normal_data[b][1][y][x]
-                    z = normal_data[b][2][y][x]
+                    x = normal_data[b][0.long()][y][x]
+                    y = normal_data[b][1.long()][y][x]
+                    z = normal_data[b][2.long()][y][x]
                     print(V.shape)
                     print(x.shape)
                     Loss += torch.dot(V,N)
